+++ src/Client/Contracts/Client.php
@@ -32,6 +32,7 @@ interface Client
      *
      * @param string $method
      * @param string $uri
+     * @param array $formData
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -39,12 +40,13 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function request(string $method, string $uri, array $options = []) : ResponseContract;
+    public function request(string $method, string $uri, array $formData = [], array $options = []): ResponseContract;
 
     /**
      * Create and send an GET HTTP request.
      *
      * @param string $uri
+     * @param array $formData
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -52,12 +54,13 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function get(string $uri, array $options = []): ResponseContract;
+    public function get(string $uri, array $formData = [], array $options = []): ResponseContract;
 
     /**
      * Create and send an POST HTTP request.
      *
      * @param string $uri
+     * @param array $formData
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -65,12 +68,13 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function post(string $uri, array $options = []): ResponseContract;
+    public function post(string $uri, array $formData = [], array $options = []): ResponseContract;
 
     /**
      * Create and send an PUT HTTP request.
      *
      * @param string $uri
+     * @param array $formData
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -78,12 +82,13 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function put(string $uri, array $options = []): ResponseContract;
+    public function put(string $uri, array $formData = [], array $options = []): ResponseContract;
 
     /**
      * Create and send an DELETE HTTP request.
      *
      * @param string $uri
+     * @param array $formData
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -91,7 +96,7 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function delete(string $uri, array $options = []): ResponseContract;
+    public function delete(string $uri, array $formData = [], array $options = []): ResponseContract;
 
     /**
      * Write header value.